Removing unwanted repositories and tests
